python-envoy/requirement.txt
grpcio-tools
logging
print_function
futures